requirements.txt
cycler==0.10.0
decorator==4.4.2
deskew==0.10.2
imageio==2.9.0
kiwisolver==1.2.0
matplotlib==3.3.0
networkx==2.4
numpy==1.19.1
opencv-python==4.3.0.36
pandas==1.1.0
pdf2image==1.13.1
Pillow==7.2.0
pyparsing==2.4.7
pytesseract==0.3.4
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
scikit-image==0.17.2
scipy==1.5.2
six==1.15.0
tifffile==2020.7.24
xlrd==1.2.0
xlwt==1.3.0
yaspin==1.0.0
